requirements.txt
pygtail>=0.10.0
elasticsearch>=6.0.0,<7.0.0
-e git+https://github.com/metwork-framework/mflog.git#egg=mflog
python-dateutil>=2.8.0
pytz>=2018.9
